fix(geometry): split a concave quad across an interior diagonal, not an outside one
Defect: `triangulate_polygon`'s `n == 4` fast path returned `[0,1,2, 0,2,3]`
unconditionally, with NO convexity test, sitting immediately ahead of the arm
that does test convexity. A fan is only valid from a vertex that sees the whole
ring, so on a concave quad the 0-2 split crosses outside the polygon.

Triggering input: the dart `[(0,0), (2,2), (4,0), (2,10)]`, reflex at vertex 1.
Twice the ring's signed area is +32.0, while `tri(0,1,2)` is -8.0 and
`tri(0,2,3)` is +40.0: the first triangle is wound BACKWARDS and lies entirely
outside the polygon, and the second covers that outside region twice. Reached
from `extrusion.rs:72` (every extruded profile cap, via
`triangulate_polygon_with_holes`), `processors/brep/faceted.rs:214`,
`processors/advanced_face/surfaces.rs:111` and `processors/sectioned.rs:558`.

Fix: pick the diagonal by orientation. `c1 = orient(0,1,2)` is the convexity of
vertex 1 and `c3 = orient(2,3,0)` that of vertex 3, and `c1 + c3` is twice the
ring's own signed area. A simple quad has at most one reflex vertex and the only
interior diagonal is the one incident to it, so agreeing signs mean 0-2 is
interior and disagreeing signs mean 1-3 is. A collinear vertex scores 0, which
agrees with anything and keeps the old split.

Kept as an orientation test rather than deleting the arm and letting
`n <= 8 && is_convex` handle quads, because this is the hot path for every face
in the pipeline. Medians of interleaved release runs of 2,000,000 calls on a
convex quad: 21.1 ns/call for the old untested arm (n=7), 25.5 for the delete
(n=4), 20.0 for the two cross products (n=3, inside run-to-run noise). Convex
quads keep byte-identical output under either option.

Evidence: three tests. The CCW dart and a CW dart ROTATED so its reflex vertex
lands on index 1 both assert every emitted triangle carries the ring's winding
sign and that the two areas sum to the ring's exactly (no double cover).
Mutation: restoring the unconditional `[0,1,2, 0,2,3]` fails both. The rotation
is load-bearing and was found by mutation: simply reversing the CCW dart puts
the reflex vertex on index 2, where the hard-coded diagonal happens to be
correct, and that fixture passed with the defect present. The third test is a
control, not a regression test: it pins that a convex quad and a
collinear-vertex quad still emit exactly `[0,1,2, 0,2,3]`, so it passes under
the mutation by design.

`cargo test -p ifc-lite-geometry` 1321 green, `cargo test -p ifc-lite-processing`
539 green including the `mesh_determinism` manifest pair, so no snapshot or
manifest re-pin. `cargo clippy -p ifc-lite-geometry --all-targets -D warnings`
clean.

The test module moves to a sibling `triangulation_tests.rs` (house `#[path]`
pattern): the new cases pushed the file to 914 lines against its recorded
ratchet budget of 782. The allowlist row is left at 782 rather than tightened to
the new 530, to keep this diff off a file parallel work also edits.
